Fix a bug giving wrong birth times in dimension 0
When computing the persistence pairs in dimension 0, and using a filtration such as sum of vertex weights, sometimes the birth time was wrong. 

The reason for this is that dset.link(u, v) was being called to late. On line 539 a test is done to see if u has been merged into v or if v has been merged into u. However, it always finds that dset.find(u)==u as the function dset.link(u, v) has not yet been called.

So I have moved dset.link(u, v) to line 531, which I think fixes the problem.